add board screen
