+++ App.js
@@ -1,113 +1,10 @@
 import { StatusBar } from 'expo-status-bar';
 import react from 'react';
-import { FlatList, StyleSheet, Text, TouchableOpacity, View, Image } from 'react-native';
+import { StyleSheet, Text, View, Image } from 'react-native';
+import BoardScreen from './src/screens/BoardScreen';
 
-const delay = ms => new Promise(res => setTimeout(res, ms));
-
-export default function App() {
+const App = () => {
   const [ notification, setNotification ] = react.useState("Player X to start!")
-  const [ refresh, setRefresh ] = react.useState(false)
-  const [ currentPlayer, setCurrentPlayer ] = react.useState("X")
-  const [ lock, setLock ] = react.useState(false)
-  const [ board, setBoard ] = react.useState(
-    [
-      " ", " ", " ",
-      " ", " ", " ",
-      " ", " ", " "
-    ]
-  )
-
-  const pressField = (index) => {
-    let newBoard = board
-    if (newBoard[index] === " " && !lock) {
-      if (currentPlayer == "X") {
-        newBoard[index] = currentPlayer
-        setCurrentPlayer("O")
-        setNotification("Player O to move!")
-      } else {
-        newBoard[index] = currentPlayer
-        setCurrentPlayer("X")
-        setNotification("Player X to move!")
-      }
-      setBoard(newBoard)
-      setRefresh(!refresh)
-
-      if (checkIfPlayerWon()) return
-
-      validClearBoard(board[index])
-    }
-  }
-
-  const checkIfPlayerWon = () => {
-    let playerWon = false
-    if (board[0] == board[1] && board[1] == board[2] && board[0] != " ") {
-      playerWon = true
-      playWon(board[0])
-    } else if (board[3] == board[4] && board[4] == board[5] && board[3] != " ") {
-      playerWon = true
-      playWon(board[3])
-    } else if (board[6] == board[7] && board[7] == board[8] && board[6] != " ") {
-      playerWon = true
-      playWon(board[6])
-    } else if (board[0] == board[4] && board[4] == board[8] && board[0] != " ") {
-      playerWon = true
-      playWon(board[0])
-    } else if (board[2] == board[4] && board[4] == board[6] && board[2] != " ") {
-      playerWon = true
-      playWon(board[2])
-    } else if (board[0] == board[3] && board[3] == board[6] && board[0] != " ") {
-      playerWon = true
-      playWon(board[0])
-    } else if (board[1] == board[4] && board[4] == board[7] && board[1] != " ") {
-      playerWon = true
-      playWon(board[1])
-    } else if (board[2] == board[5] && board[5] == board[8] && board[2] != " ") {
-      playerWon = true
-      playWon(board[2])
-    }
-    return playerWon
-  }
-
-  const playWon = async(player) => {
-    setNotification("Player " + player + " won!")
-    setLock(true)
-    await delay(2000)
-    resetGame(player)
-  }
-
-  const validClearBoard = (player) => {
-    for (const cell of board) {
-      if (cell == " ") return
-    }
-
-    gameOver(player)
-  }
-
-  const gameOver = async(player) => {
-    setNotification("Game over!")
-    setLock(true)
-    await delay(2000)
-
-    resetGame(player)
-  }
-
-  const resetGame = (player) => {
-    clearBoard()
-    setLock(false)
-
-    if (player == "X") setNotification("Player O to move!")
-    else setNotification("Player X to move!")
-  }
-
-  const clearBoard = () => {
-    setBoard(
-      [
-        " ", " ", " ",
-        " ", " ", " ",
-        " ", " ", " "
-      ]
-    )
-  }
 
   return (
     <View style={styles.container}>
@@ -118,24 +15,7 @@ export default function App() {
       <StatusBar style="auto" />
       <Text style={styles.txt1}>TicTacToe</Text>
       <Text style={styles.txt2}>{notification}</Text>
-      <View style={styles.flatlistContainer}>
-        <Image
-          source={require('./assets/bg.png')}
-          style={styles.image}
-        />
-        <FlatList
-          style={styles.list}
-          data={board}
-          numColumns={3}
-          refreshing={true}
-          extraData={refresh}
-          renderItem={({item, index}) =>
-            <TouchableOpacity style={styles.square} onPress={() => pressField(index)}>
-              <Text style={styles.txtXO}>{item}</Text>
-            </TouchableOpacity>
-          }
-        />
-      </View>
+      <BoardScreen setNotification={setNotification} />
     </View>
   );
 }
@@ -191,3 +71,5 @@ const styles = StyleSheet.create({
     height: '100%'
   }
 });
+
+export default App;
